move file to specific folder and update code with ui design
